+++ packages/database/constants.ts
@@ -35,61 +35,6 @@ export const allowedMimeTypes = [
 	'application/x-mobipocket-ebook',
 ]
 
-export const popularCurrencies = [
-	'USD', // United States Dollar
-	'EUR', // Euro
-	'JPY', // Japanese Yen
-	'GBP', // Pound Sterling
-	'CHF', // Swiss Franc
-	'CNY', // Chinese Renminbi (RMB)
-	'AUD', // Australian Dollar
-	'CAD', // Canadian Dollar
-	'HKD', // Hong Kong Dollar
-	'SGD', // Singapore Dollar
-	'INR', // Indian Rupee
-	'MXN', // Mexican Peso
-	'RUB', // Russian Ruble
-	'BRL', // Brazilian Real
-	'TRY', // Turkish Lira
-	'KRW', // South Korean Won
-	'ZAR', // South African Rand
-	'ARS', // Argentine Peso
-	'CLP', // Chilean Peso
-	'COP', // Colombian Peso
-	'PEN', // Peruvian Sol
-	'UYU', // Uruguayan Peso
-	'VEF', // Venezuelan Bolivar (note: highly volatile and subject to frequent changes)
-	'PHP', // Philippine Peso
-	'THB', // Thai Baht
-	'IDR', // Indonesian Rupiah
-	'MYR', // Malaysian Ringgit
-]
-
-export const devUser1 = {
-	pk: '86a82cab18b293f53cbaaae8cdcbee3f7ec427fdf9f9c933db77800bb5ef38a0',
-	sk: '5c81bffa8303bbd7726d6a5a1170f3ee46de2addabefd6a735845166af01f5c0',
-}
-
-export const devUser2 = {
-	pk: 'd943e96d62695b318a9c0658a3bd3fafaaf441a069d8bfd04dc9ff39c69cc782',
-	sk: '08a475839723c79f2993ad000289670eb737d34bc9d72d43128f898713fc3fb3',
-}
-
-export const devUser3 = {
-	pk: '2edec1b799cd2f41f70a5ff0edc10d2260a57d62f39072aab4eb8174b7ca912a',
-	sk: 'e61ae5a4f505026e3d2b5aeba82c748b6b799346a1e98e266d7252cddb8f502b',
-}
-
-export const devUser4 = {
-	pk: 'f47121cd783802e6d4879e63233b54aff54e6788ea9ef568cec0259cc60fe286',
-	sk: 'beb8f6777d4379ac60b01d91fa84456bb23a2ef6b083f557b9ede311ae1ede53',
-}
-
-export const devUser5 = {
-	pk: '96c727f4d1ea18a80d03621520ebfe3c9be1387033009a4f5b65959d09222eec',
-	sk: 'ee40a2dc441238f241d1728af9507147e9b5ed18c1c61d84876d4f2502c044b3',
-}
-
 export type ObjectValues<T> = T[keyof T]
 
 export const META_SCOPES = {
